use correct type

+++ src/Model/Table/CronjobsTable.php
@@ -212,7 +212,7 @@ class CronjobsTable extends AppTable
     {
 
         if (empty($this->cronjobRunDay)) {
-            $this->cronjobRunDay = Configure::read('app.timeHelper')->getTimeObjectUTC(date(Configure::read('DateFormat.DatabaseWithTimeAlt')))->toUnixString();
+            $this->cronjobRunDay = (int) Configure::read('app.timeHelper')->getTimeObjectUTC(date(Configure::read('DateFormat.DatabaseWithTimeAlt')))->toUnixString();
         }
 
         $this->CronjobLogs->deleteOldLogs($this->cronjobRunDay);
